requirements.txt
# Core Framework
langchain>=0.2.0
langchain-community>=0.2.0

# Gemini Integration (FREE)
langchain-google-genai>=2.0.0
google-generativeai>=0.4.0

# Vector Store (lightweight, local)
numpy>=1.26.0

# PDF Support (NEW!)
pypdf>=4.0.0

# Web UI
streamlit>=1.30.0

# Environment
python-dotenv>=1.0.0

# Data Validation
pydantic==2.7.4

# Optional (advanced file parsing)
pdfplumber>=0.11.0  # Alternative to pypdf (more advanced)